finverify/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finverify"
version = "0.1.0"
description = "Automatically verify financial reports from PDF/Excel/CSV using formulas"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "FinVerify", email = "noreply@example.com" }]
dependencies = [
  "pandas>=2.0.0",
  "openpyxl>=3.1.0",
  "xlrd>=2.0.1",
  "pdfplumber>=0.11.0",
  "pytesseract>=0.3.10",
  "pdf2image>=1.17.0",
  "Pillow>=10.0.0",
]

[project.scripts]
finverify = "finverify.cli:main"

